+++ sample-application/src/main/java/com/alternativepayments/controller/TransactionController.java
@@ -1,7 +1,5 @@
 package com.alternativepayments.controller;
 
-import java.math.BigDecimal;
-
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -36,7 +34,7 @@ public class TransactionController {
         Customer customer = new Customer.Builder("John", "Doe", "john@doe.com", "DE").build();
         Payment payment = new Payment.Builder("SEPA", "John Doe").iban("BE88271080782541").build();
 
-        Transaction sepaTransaction = new Transaction.Builder(payment, null, BigDecimal.valueOf(500), "EUR")
+        Transaction sepaTransaction = new Transaction.Builder(payment, null, 500, "EUR")
                 .customer(customer).build();
         Transaction createdTransaction = alternativePaymentClient.create(sepaTransaction, Transaction.API_ENDPOINT,
                 Transaction.class);
@@ -57,7 +55,7 @@ public class TransactionController {
         RedirectUrls redirectUrls = new RedirectUrls("http://plugins.alternativepayments.com/message/success.html",
                 "http://plugins.alternativepayments.com/message/failure.html");
 
-        Transaction mistercashTransaction = new Transaction.Builder(payment, null, BigDecimal.valueOf(500), "EUR")
+        Transaction mistercashTransaction = new Transaction.Builder(payment, null, 500, "EUR")
                 .redirectUrls(redirectUrls).customer(customer).build();
         Transaction createdTransaction = alternativePaymentClient.create(mistercashTransaction,
                 Transaction.API_ENDPOINT, Transaction.class);
@@ -78,7 +76,7 @@ public class TransactionController {
         RedirectUrls redirectUrls = new RedirectUrls("http://plugins.alternativepayments.com/message/success.html",
                 "http://plugins.alternativepayments.com/message/failure.html");
 
-        Transaction idealTransaction = new Transaction.Builder(payment, null, BigDecimal.valueOf(500), "EUR")
+        Transaction idealTransaction = new Transaction.Builder(payment, null, 500, "EUR")
                 .redirectUrls(redirectUrls).customer(customer).build();
         Transaction createdTransaction = alternativePaymentClient.create(idealTransaction,
                 Transaction.API_ENDPOINT, Transaction.class);
@@ -101,7 +99,7 @@ public class TransactionController {
         RedirectUrls redirectUrls = new RedirectUrls("http://plugins.alternativepayments.com/message/success.html",
                 "http://plugins.alternativepayments.com/message/failure.html");
 
-        Transaction brazilPayTransaction = new Transaction.Builder(payment, null, BigDecimal.valueOf(100), "EUR")
+        Transaction brazilPayTransaction = new Transaction.Builder(payment, null, 100, "EUR")
                 .redirectUrls(redirectUrls).customer(customer).build();
         Transaction createdTransaction = alternativePaymentClient.create(brazilPayTransaction,
                 Transaction.API_ENDPOINT, Transaction.class);
@@ -121,7 +119,7 @@ public class TransactionController {
         Payment payment = new Payment.Builder("CreditCard", "John Doe").creditCardNumber("4111111111111111")
                 .creditCardType("visa").expirationMonth(12).expirationYear(2019).cvv2(222).build();
 
-        Transaction creditCardTransaction = new Transaction.Builder(payment, null, BigDecimal.valueOf(100), "EUR")
+        Transaction creditCardTransaction = new Transaction.Builder(payment, null, 100, "EUR")
                 .customer(customer).build();
 
         Transaction createdTransaction = alternativePaymentClient.create(creditCardTransaction,
@@ -142,7 +140,7 @@ public class TransactionController {
         Payment payment = new Payment.Builder("Teleingreso", "John Doe").build();
         RedirectUrls redirectUrls = new RedirectUrls("http://plugins.alternativepayments.com/message/success.html",
                 "http://plugins.alternativepayments.com/message/failure.html");
-        Transaction teleingresoTransaction = new Transaction.Builder(payment, null, BigDecimal.valueOf(100), "EUR")
+        Transaction teleingresoTransaction = new Transaction.Builder(payment, null, 100, "EUR")
                 .redirectUrls(redirectUrls).customer(customer).build();
 
         Transaction createdTransaction = alternativePaymentClient.create(teleingresoTransaction,
@@ -163,7 +161,7 @@ public class TransactionController {
         Payment payment = new Payment.Builder("Teleingreso", "John Doe").build();
         RedirectUrls redirectUrls = new RedirectUrls("http://plugins.alternativepayments.com/message/success.html",
                 "http://plugins.alternativepayments.com/message/failure.html");
-        Transaction safetyPayTransaction = new Transaction.Builder(payment, null, BigDecimal.valueOf(100), "EUR")
+        Transaction safetyPayTransaction = new Transaction.Builder(payment, null, 100, "EUR")
                 .redirectUrls(redirectUrls).customer(customer).build();
 
         Transaction createdTransaction = alternativePaymentClient.create(safetyPayTransaction,
@@ -185,7 +183,7 @@ public class TransactionController {
         Payment payment = new Payment.Builder("POLi", "John Doe").build();
         RedirectUrls redirectUrls = new RedirectUrls("http://plugins.alternativepayments.com/message/success.html",
                 "http://plugins.alternativepayments.com/message/failure.html");
-        Transaction poliTransaction = new Transaction.Builder(payment, null, BigDecimal.valueOf(100), "EUR")
+        Transaction poliTransaction = new Transaction.Builder(payment, null, 100, "EUR")
                 .redirectUrls(redirectUrls).customer(customer).build();
 
         Transaction createdTransaction = alternativePaymentClient.create(poliTransaction,
@@ -207,7 +205,7 @@ public class TransactionController {
         Payment payment = new Payment.Builder("TrustPay", "John Doe").build();
         RedirectUrls redirectUrls = new RedirectUrls("http://plugins.alternativepayments.com/message/success.html",
                 "http://plugins.alternativepayments.com/message/failure.html");
-        Transaction trustPayTransaction = new Transaction.Builder(payment, null, BigDecimal.valueOf(100), "EUR")
+        Transaction trustPayTransaction = new Transaction.Builder(payment, null, 100, "EUR")
                 .redirectUrls(redirectUrls).customer(customer).build();
 
         Transaction createdTransaction = alternativePaymentClient.create(trustPayTransaction,
@@ -229,7 +227,7 @@ public class TransactionController {
         Payment payment = new Payment.Builder("directpay", "John Doe").build();
         RedirectUrls redirectUrls = new RedirectUrls("http://plugins.alternativepayments.com/message/success.html",
                 "http://plugins.alternativepayments.com/message/failure.html");
-        Transaction directPayTransaction = new Transaction.Builder(payment, null, BigDecimal.valueOf(100), "EUR")
+        Transaction directPayTransaction = new Transaction.Builder(payment, null, 100, "EUR")
                 .redirectUrls(redirectUrls).customer(customer).build();
 
         Transaction createdTransaction = alternativePaymentClient.create(directPayTransaction,
@@ -251,7 +249,7 @@ public class TransactionController {
         Payment payment = new Payment.Builder("directpaymax", "John Doe").bankCode("POSTBANK").build();
         RedirectUrls redirectUrls = new RedirectUrls("http://plugins.alternativepayments.com/message/success.html",
                 "http://plugins.alternativepayments.com/message/failure.html");
-        Transaction directPayMaxTransaction = new Transaction.Builder(payment, null, BigDecimal.valueOf(100), "EUR")
+        Transaction directPayMaxTransaction = new Transaction.Builder(payment, null, 100, "EUR")
                 .redirectUrls(redirectUrls).customer(customer).build();
 
         Transaction createdTransaction = alternativePaymentClient.create(directPayMaxTransaction,
